refactor analysis classes - wip
